Convert Agent/Skill/Script edit forms from inline Card to Modal overlay
Replace inline Card-based create/edit forms with proper Modal dialogs
for Agents, Skills, and Scripts, matching the existing Task creation
modal pattern for consistent UX.

- AgentFormModal: swap Card wrapper for Modal with Header/Body/Footer
- SkillFormModal: extract inline form from SkillList into new component
- ScriptFormModal: extract inline form from ScriptList into new component
- SkillListUtils: extract SkillFormData type and helpers from SkillList
